Lower rubocop-rails version to v2.12
Jumping all the way to v2.27 resulted in way more violations in root-server than I was comfortable fixing right now. I just wanted to get rid of the pesky error message, which says it'll stop with v2.9.

I originally tried to go as low as v2.9.0, but I ran into an issue with the Rails/ContentTag cop trying to replace our content_tag usage in a breaking way... It looks like this was fixed in v2.12.0: rubocop/rubocop-rails#526.