Relax rubocop-rails pinning
This allows consumings apps to bump rubocop-rails beyond 2.31 if they wish, but they are not forced to do so.

rubocop-rails 2.32 includes an important update that allows Rails rubocop rules to apply correctly to files in nested folder structures, which is common in apps that use engines.

When a consuming app upgrades to 2.32 it likely will cause a sudden influx in Rails rubocop violations.